refactor: изменение названий подмодулей
