+++ src/external/learning_analytics/assessment_diagnostics_submodule/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 class KnowledgeModuleConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'src.external.learning_analytics.knowledge_module'
+    name = 'src.external.learning_analytics.assessment_diagnostics_submodule'
     label = 'learning_analytics_knowledge_module'
