+++ src/external/learning_analytics/predictive_analytics_submodule/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 class TrainingModuleConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'src.external.learning_analytics.training_module'
+    name = 'src.external.learning_analytics.predictive_analytics_submodule'
     label = 'learning_analytics_training_module'
