+++ src/external/learning_analytics/visualization_reporting_submodule/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 class MonitoringModuleConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'src.external.learning_analytics.monitoring_module'
+    name = 'src.external.learning_analytics.visualization_reporting_submodule'
     label = 'learning_analytics_monitoring_module'
